requirements.txt
fastapi==0.119.0
langchain_community==0.3.31
langchain_core==0.3.79
langchain_mistralai==0.2.12
python-dotenv==1.1.1
uvicorn==0.37.0
